delete undefined

+++ finalProject.py
@@ -720,7 +720,8 @@ def deleteSingleReport(report_id):
     reportToDelete = session.query(Report).filter_by(id=report_id).one()
     filename = reportToDelete.report_picture
     if request.method == 'POST':
-        os.remove("%s%s" % (upload_folder, filename))
+        if ("%s%s" % (upload_folder, filename)):
+            os.remove("%s%s" % (upload_folder, filename))
         session.delete(reportToDelete)
         session.commit()
         flash("report was deleted !")
